+++ framework/i18n/src/test/java/is/codion/framework/i18n/FrameworkMessagesTest.java
@@ -37,6 +37,7 @@ public class FrameworkMessagesTest {
 		FrameworkMessages.updateMnemonic();
 		FrameworkMessages.updateTip();
 		FrameworkMessages.edit();
+		FrameworkMessages.editMnemonic();
 		FrameworkMessages.editSelectedTip();
 		FrameworkMessages.delete();
 		FrameworkMessages.deleteMnemonic();